+++ src/messageExtension/templates/src/app/{messageExtensionName}/{messageExtensionClassName}.ts
@@ -12,16 +12,58 @@ const log = debug("msteams");
 export default class <%= messageExtensionClassName%> implements IMessagingExtensionMiddlewareProcessor {
 
     public async onQuery(context: TurnContext, query: MessagingExtensionQuery): Promise<MessagingExtensionResult> {
-        const card = CardFactory.heroCard("Test", "Test", [`https://${process.env.HOSTNAME}/assets/icon.png`]);
+        const card = CardFactory.adaptiveCard(
+            {
+                type: "AdaptiveCard",
+                body: [
+                    {
+                        type: "TextBlock",
+                        size: "Large",
+                        text: "Headline"
+                    },
+                    {
+                        type: "TextBlock",
+                        text: "Description"
+                    },
+                    {
+                        type: "Image",
+                        url: `https://${process.env.HOSTNAME}/assets/icon.png`
+                    }
+                ],
+                actions: [
+                    {
+                        type: "Action.Submit",
+                        title: "More details",
+                        data: {
+                            action: "moreDetails",
+                            id: "1234-5678"
+                        }
+                    }
+                ],
+                $schema: "http://adaptivecards.io/schemas/adaptive-card.json",
+                version: "1.0"
+            });
+        const preview = {
+            contentType: "application/vnd.microsoft.card.thumbnail",
+            content: {
+                title: "Headline",
+                text: "Description",
+                images: [
+                    {
+                        url: "`https://${process.env.HOSTNAME}/assets/icon.png`
+                    }
+                ]
+            }
+        };
 
         if (query.parameters && query.parameters[0] && query.parameters[0].name === "initialRun") {
             // initial run
 
             return Promise.resolve({
                 type: "result",
-                attachmentLayout: "grid",
+                attachmentLayout: "list",
                 attachments: [
-                    card
+                    { ...card, preview }
                 ]
             } as MessagingExtensionResult);
         } else {
@@ -30,7 +72,7 @@ export default class <%= messageExtensionClassName%> implements IMessagingExtens
                 type: "result",
                 attachmentLayout: "list",
                 attachments: [
-                    card
+                    { ...card, preview }
                 ]
             } as MessagingExtensionResult);
         }
@@ -48,6 +90,14 @@ export default class <%= messageExtensionClassName%> implements IMessagingExtens
         // take care of the setting returned from the dialog, with the value stored in state
         const setting = context.activity.value.state;
         log(`New setting: ${setting}`);
+        return Promise.resolve();
+    }
+
+    public async onCardButtonClicked(context: TurnContext, value: any): Promise<void> {
+        // Handle the Action.Submit action on the adaptive card
+        if (value.action === "moreDetails") {
+            log(`I got this ${value.id}`);
+        }
         return Promise.resolve();
     }